+++ mobile/android/android-components/components/browser/session/src/test/java/mozilla/components/browser/session/SessionManagerMigrationTest.kt
@@ -13,7 +13,6 @@ import mozilla.components.concept.engine.EngineSession
 import mozilla.components.concept.engine.EngineSessionState
 import mozilla.components.concept.engine.HitResult
 import mozilla.components.concept.engine.content.blocking.Tracker
-import mozilla.components.concept.engine.prompt.PromptRequest
 import mozilla.components.support.base.observer.Consumable
 import mozilla.components.support.test.mock
 import org.junit.Assert.assertEquals
@@ -24,7 +23,6 @@ import org.junit.Assert.assertSame
 import org.junit.Assert.assertTrue
 import org.junit.Test
 import org.mockito.Mockito.doReturn
-import org.mockito.Mockito.`when`
 
 /**
  * This test suite validates that calls on [SessionManager] update [BrowserStore] to create a matching state.
@@ -797,32 +795,6 @@ class SessionManagerMigrationTest {
 
         store.state.findTab("session2")!!.also { tab ->
             assertEquals(engineSession, tab.engineState.engineSession)
-        }
-    }
-
-    @Test
-    fun `Adding a prompt request`() {
-        val store = BrowserStore()
-        val manager = SessionManager(engine = mock(), store = store)
-
-        val session = Session(id = "session", initialUrl = "https://www.mozilla.org")
-        manager.add(session)
-
-        assertNull(session.promptRequest.peek())
-        assertNull(store.state.findTab("session")!!.content.promptRequest)
-
-        val promptRequest: PromptRequest = mock()
-        session.promptRequest = Consumable.from(promptRequest)
-
-        assertEquals(promptRequest, session.promptRequest.peek())
-        store.state.findTab("session")!!.also { tab ->
-            assertNotNull(tab.content.promptRequest)
-            assertSame(promptRequest, tab.content.promptRequest)
-        }
-
-        session.promptRequest.consume { true }
-        store.state.findTab("session")!!.also { tab ->
-            assertNull(tab.content.promptRequest)
         }
     }